+++ jib-maven-plugin/src/test/java/com/google/cloud/tools/jib/maven/BuildDockerMojoIntegrationTest.java
@@ -56,7 +56,7 @@ public class BuildDockerMojoIntegrationTest {
   @Test
   public void testExecute_simple() throws VerificationException, IOException, InterruptedException {
     Assert.assertEquals(
-        "Hello, world\n",
+        "Hello, world. An argument.\n",
         buildToDockerDaemonAndRun(
             simpleTestProject.getProjectRoot(),
             "gcr.io/jib-integration-testing/simpleimage:maven"));