+++ jib-maven-plugin/CHANGELOG.md
@@ -15,11 +15,16 @@ All notable changes to this project will be documented in this file.
 - Warns if build may not be reproducible ([#245](https://github.com/GoogleContainerTools/jib/pull/245))
 - `jib:buildDocker` maven goal to build straight to Docker daemon ([#266](https://github.com/GoogleContainerTools/jib/pull/266))
 - `mainClass` is inferred by searching through class files if configuration is missing ([#278](https://github.com/GoogleContainerTools/jib/pull/278))
+- Can now specify target image with `-Dimage` ([#328](https://github.com/GoogleContainerTools/jib/issues/328))
+- `args` parameter to define default main args ([#346](https://github.com/GoogleContainerTools/jib/issues/346))
 
 ### Changed
 - Removed `enableReproducibleBuilds` parameter - application layers will always be reproducible ([#245](https://github.com/GoogleContainerTools/jib/pull/245))
 - Changed configuration schema to be more like configuration for `jib-gradle-plugin` - NOT compatible with prior versions of `jib-maven-plugin` ([#212](https://github.com/GoogleContainerTools/jib/issues/212))
 - `jib:dockercontext` has been changed to `jib:dockerContext` ([#321](https://github.com/GoogleContainerTools/jib/issues/321))
+
+### Fixed
+- Directories in resources are added to classes layer ([#318](https://github.com/GoogleContainerTools/jib/issues/318))
 
 ## 0.1.7
 ### Fixed